requirements.txt
ultralytics==8.2.0
deep-sort-realtime==1.3.2
deepface==0.0.93
umap-learn==0.5.6
streamlit==1.35.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.0
scikit-learn==1.4.2
Pillow==10.3.0
tqdm==4.66.4
torch>=2.0.0
torchvision>=0.15.0
tf-keras==2.16.0
tensorflow==2.16.1
plotly==5.22.0
scipy==1.13.0
